pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=23.7.0",
]

[project]
name = "backend_pedidos-now"
version = "0.1.0"
description = ""
authors = [
    {name = "lamix1", email = "camilacristina328santoss@gmail.com"},
]
dependencies = [
    "django>=4.2.4",
    "djangorestframework>=3.14.0",
    "django-extensions>=3.2.3",
    "pydotplus>=2.0.2",
    "django-cors-headers>=4.2.0",
    "drf-spectacular>=0.26.5",
    "django-filter>=23.3",
    "djangorestframework-simplejwt>=5.3.0",
    "python-magic>=0.4.27",
    "Pillow>=10.0.1",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
